requirements.txt
# Django Core
Django>=5.0,<6.0
django-environ>=0.11.2
django-extensions>=3.2.3

# Database
django-q2>=1.6.2

# Authentication
django-allauth>=0.61.1

# AI & ML (All Free, Local)
sentence-transformers>=2.2.2
torch>=2.1.0
spacy>=3.7.2
httpx>=0.26.0  # For Ollama API calls

# PDF Processing
pdfplumber>=0.10.3
PyMuPDF>=1.23.8
pytesseract>=0.3.10
Pillow>=10.2.0

# PDF Report Generation
WeasyPrint>=60.2
Jinja2>=3.1.2

# NLP
nltk>=3.8.1

# Utilities
python-magic>=0.4.27
numpy>=1.26.3
scikit-learn>=1.4.0

# Development
django-debug-toolbar>=4.2.0
ipython>=8.20.0

# Testing
pytest>=7.4.4
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
